+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## vNEXT
-- Check if a voucher exists at a given address. (#42)
+- Add `predictVoucher(..)` & `isVoucher(..)` functions. (#42)
 - Generate UML class diagram for contracts. (#41)
 - Generate Solidity documentation. (#40)
 - Add Bellecour poco address to config and harmonize deployment artifacts. (#39)